requirements.txt
django-bootstrap-toolkit==2.15.0
django-bootstrap3==11.0.0
gunicorn==19.9.0
Pillow==5.3.0
pytz==2018.7
whitenoise==4.1.2
